% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpgrow.R
\name{lgmPost}
\alias{lgmPost}
\title{Run a Bayesian mixed effects model for by-subject random effects with an independent Gaussian prior}
\usage{
lgmPost(y, X, Z, subjects, niter, nburn, nthin)
}
\arguments{
\item{y}{An \emph{N x 1} response (of subject-measure cases)}

\item{X}{Fixed effects design matrix}

\item{Z}{Random effects design matrix.  Assumed grouped by \code{subjects}}

\item{subjects}{An \emph{N x 1} set of subject identifiers}

\item{niter}{The number of MCMC iterations}

\item{nburn}{The number of MCMC burn-in iterations to discard}

\item{nthin}{The step increment of MCMC samples to return
The rate parameter is set of \code{1}.}
}
\value{
res A list object containing MCMC runs for all model parameters.
}
\description{
An internal function to \code{\link{dpgrow}}
}
\note{
Intended as an internal function for \code{\link{dpgrow}}
}
\author{
Terrance Savitsky \email{tds151@gmail.com}
}
\seealso{
\code{\link{dpgrow}}
}

