\name{samples.ddpgrow}
\alias{samples.ddpgrow}
\title{Produce samples of MCMC output}
\usage{
  \method{samples}{ddpgrow} (object, ...)
}
\arguments{
  \item{object}{A \code{ddpgrow} object}

  \item{...}{Ignored}
}
\description{
  provides posterior sampled values for every model
  parameter of a \code{ddpgrow} object
}

