% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddpgrow.R
\name{summary.ddpgrow}
\alias{summary.ddpgrow}
\title{S3 functions of dpgrow}
\usage{
\method{summary}{ddpgrow}(object, ...)
}
\arguments{
\item{object}{A \code{ddpgrow} object}

\item{...}{Ignored}
}
\description{
produces quantile summaries for model parameters
}

