% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpgrow.R
\name{summary.dpgrow}
\alias{summary.dpgrow}
\title{S3 functions of dpgrow}
\usage{
\method{summary}{dpgrow}(object, ...)
}
\arguments{
\item{object}{A \code{dpgrow} object}

\item{...}{Ignored}
}
\description{
produces quantile summaries for model parameters
}

