% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpgrowmm.R
\name{summary.dpgrowmm}
\alias{summary.dpgrowmm}
\title{S3 functions of dpgrowmm}
\usage{
\method{summary}{dpgrowmm}(object, ...)
}
\arguments{
\item{object}{A \code{dpgrowmm} object}

\item{...}{Ignored}
}
\description{
produces quantile summaries for model parameters
for an \code{dgprowmm} object.
}

