% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpgrow.R
\name{samples.dpgrow}
\alias{samples.dpgrow}
\title{Produce samples of MCMC output}
\usage{
\method{samples}{dpgrow}(object, ...)
}
\arguments{
\item{object}{A \code{dpgrow} object}

\item{...}{Ignored}
}
\description{
provides posterior sampled values for every model parameter of a
\code{dpgrow} object
}

