\name{samples.dpgrowmult}
\alias{samples.dpgrowmult}
\title{Produce samples of MCMC output}
\usage{
\method{samples}{dpgrowmult}(object, ...)
}
\arguments{
  \item{object}{A \code{dpgrowmult} object}

  \item{...}{Ignored}
}
\description{
provides posterior sampled values for every model parameter
of a \code{dpgrowmult} object
}

