\name{summary.dpgrowmult}
\alias{summary.dpgrowmult}
\title{S3 functions of dpgrowmult}
\usage{
  \method{summary}{dpgrowmult} (object, ...)
}
\arguments{
  \item{object}{A \code{dpgrowmult} object}

  \item{...}{Ignored}
}
\description{
  produces quantile summaries for model parameters for an
  \code{dgprowmult} object.
}

